chore: remove async mode from provider
Use react-promise-status instead.
